Add weeknightly supply chain review, starting with pre-commit pins
Source: This code has been trialed throughout several repositories in
the Cyber Domain Ontology.  I can't currently recall the original
repository where this logic was deployed, but this state of the code was
copied with a few typo fixes from CASE-Corpora.

References:
* https://github.com/casework/CASE-Corpora

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>